Removed called to feOpenLocalCluster
I removed all the calls to feOpenLocalCluster. This change is not strictly necessary, but it will help with compatibility between different versions of matlab and cost of speed reduction. Likely speed reduction will be minimal. Importantly many of these operations in fefgGet.m will be replaced by calls for the multiway array (Phi, tensor) in life 2.0.